move is_protected into translate.py
