+++ app/tests/test_translate.py
@@ -14,7 +14,7 @@ from app.db import model
 from app.external.rawls_entity_model import Entity
 from app.server import requestutils
 from app.tests import testutils
-from app.translate import FILETYPE_TRANSLATORS, validate_import_url
+from app.translate import FILETYPE_TRANSLATORS, validate_import_url, is_protected
 from app.translators import Translator
 from app.util.exceptions import InvalidPathException
 
@@ -370,3 +370,17 @@ def test_validate_import_url(import_url, is_valid, file_type_translator):
             validate_import_url(import_url=import_url, import_filetype=file_type_translator, user_info=user_info)
     else:
         assert validate_import_url(import_url=import_url, import_filetype=file_type_translator, user_info=user_info) is is_valid
+
+@pytest.mark.parametrize("import_url, protected", [
+    ("something.anvil.gi.ucsc.edu", True),
+    ("something-else.anvil.gi.ucsc.edu", True),
+    ("something.anvilproject.org", True),
+    ("something.anvil.edu", False),
+    ("something.org", False)
+])
+@pytest.mark.parametrize("file_type", ["pfb", "tdrexport"])
+def test_is_protected(import_url, protected, file_type):
+    if file_type == "pfb":
+        assert is_protected(import_netloc=import_url, import_filetype=file_type) is protected
+    else:
+        assert is_protected(import_netloc=import_url, import_filetype=file_type) is False
